math cleanup
