+++ packages/engine/src/avatar/AvatarTeleportSystem.ts
@@ -13,9 +13,8 @@ import {
   Vector3
 } from 'three'
 
-import { normalizeRange } from '@xrengine/common/src/utils/mathUtils'
-
 import checkPositionIsValid from '../common/functions/checkPositionIsValid'
+import { normalizeRange } from '../common/functions/MathFunctions'
 import { World } from '../ecs/classes/World'
 import { addComponent, defineQuery, getComponent } from '../ecs/functions/ComponentFunctions'
 import { createEntity } from '../ecs/functions/EntityFunctions'
