+++ packages/engine/src/common/functions/BitFunctions.ts
