+++ packages/engine/src/scene/functions/loaders/BoxColliderFunctions.test.ts
@@ -2,8 +2,8 @@ import assert from 'assert'
 import { Quaternion, Vector3 } from 'three'
 
 import { ComponentJson } from '@xrengine/common/src/interfaces/SceneInterface'
-import { quaternionEqualsEpsilon, vector3EqualsEpsilon } from '@xrengine/common/src/utils/mathUtils'
 
+import { quaternionEqualsEpsilon, vector3EqualsEpsilon } from '../../../../tests/util/MathTestUtils'
 import { Engine } from '../../../ecs/classes/Engine'
 import { Entity } from '../../../ecs/classes/Entity'
 import { addComponent, getComponent, hasComponent } from '../../../ecs/functions/ComponentFunctions'
